requirements.txt
aiohttp>=3.8.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ollama==0.4.9
passlib[bcrypt]
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv>=0.19.0 
python-jose
python-multipart
redis==6.2.0
requests==2.32.3
sniffio==1.3.1
sqlalchemy
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn[standard]==0.34.2
websockets==11.0.3
zstandard==0.21.0
fastapi
uvicorn
aiohttp
python-dotenv
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
redis
aioredis
